+++ tapeserver/castor/tape/tapeserver/file/OsmFileReader.cpp
@@ -15,34 +15,36 @@
  *               submit itself to any jurisdiction.
  */
 
+#include <limits>
 #include <memory>
 #include <sstream>
+#include <string>
 
+#include "castor/tape/tapeserver/drive/DriveInterface.hpp"
+#include "castor/tape/tapeserver/file/HeaderChecker.hpp"
+#include "castor/tape/tapeserver/file/OsmFileReader.hpp"
 #include "castor/tape/tapeserver/file/CtaReadSession.hpp"
-#include "castor/tape/tapeserver/file/OsmReadSession.hpp"
-#include "castor/tape/tapeserver/file/ReadSessionFactory.hpp"
-#include "common/dataStructures/LabelFormat.hpp"
+#include "castor/tape/tapeserver/file/Structures.hpp"
+#include "scheduler/RetrieveJob.hpp"
 
 namespace castor {
 namespace tape {
 namespace tapeFile {
 
-std::unique_ptr<ReadSession> ReadSessionFactory::create(tapeserver::drive::DriveInterface &drive,
-  const tapeserver::daemon::VolumeInfo &volInfo, const bool useLbp) {
-  using LabelFormat = cta::common::dataStructures::Label::Format;
-  const LabelFormat labelFormat = volInfo.labelFormat;
-  switch (labelFormat) {
-    case LabelFormat::CTA:
-      return std::make_unique<CtaReadSession>(drive, volInfo, useLbp);
-    case LabelFormat::OSM:
-      return std::make_unique<OsmReadSession>(drive, volInfo, useLbp);
-    default: {
-      std::ostringstream ossLabelFormat;
-      ossLabelFormat << std::showbase << std::internal << std::setfill('0') << std::hex << std::setw(4)
-                     << static_cast<unsigned int>(labelFormat);
-      throw TapeFormatError("In ReadSessionFactory::create(): unknown label format: " + ossLabelFormat.str());
-    }
-  }
+OsmFileReader::OsmFileReader(const std::unique_ptr<ReadSession> &rs, const cta::RetrieveJob &fileToRecall)
+  : FileReader(rs, fileToRecall) {
+}
+
+void OsmFileReader::positionByFseq(const cta::RetrieveJob &fileToRecall) {
+  throw NotImplemented("OsmFileReader::positionByFseq() needs to be implemented");
+}
+
+void OsmFileReader::positionByBlockID(const cta::RetrieveJob &fileToRecall) {
+  throw NotImplemented("OsmFileReader::positionByBlockID() needs to be implemented");
+}
+
+size_t OsmFileReader::readNextDataBlock(void *data, const size_t size) {
+  throw NotImplemented("OsmFileReader::readNextDataBlock() needs to be implemented");
 }
 
 }  // namespace tapeFile