+++ tapeserver/castor/tape/tapeserver/file/OsmReadSession.cpp
@@ -16,33 +16,21 @@
  */
 
 #include <memory>
-#include <sstream>
+#include <string>
 
-#include "castor/tape/tapeserver/file/CtaReadSession.hpp"
+#include "castor/tape/tapeserver/file/Exceptions.hpp"
+#include "castor/tape/tapeserver/file/HeaderChecker.hpp"
 #include "castor/tape/tapeserver/file/OsmReadSession.hpp"
-#include "castor/tape/tapeserver/file/ReadSessionFactory.hpp"
-#include "common/dataStructures/LabelFormat.hpp"
+#include "castor/tape/tapeserver/file/Structures.hpp"
 
 namespace castor {
 namespace tape {
 namespace tapeFile {
 
-std::unique_ptr<ReadSession> ReadSessionFactory::create(tapeserver::drive::DriveInterface &drive,
-  const tapeserver::daemon::VolumeInfo &volInfo, const bool useLbp) {
-  using LabelFormat = cta::common::dataStructures::Label::Format;
-  const LabelFormat labelFormat = volInfo.labelFormat;
-  switch (labelFormat) {
-    case LabelFormat::CTA:
-      return std::make_unique<CtaReadSession>(drive, volInfo, useLbp);
-    case LabelFormat::OSM:
-      return std::make_unique<OsmReadSession>(drive, volInfo, useLbp);
-    default: {
-      std::ostringstream ossLabelFormat;
-      ossLabelFormat << std::showbase << std::internal << std::setfill('0') << std::hex << std::setw(4)
-                     << static_cast<unsigned int>(labelFormat);
-      throw TapeFormatError("In ReadSessionFactory::create(): unknown label format: " + ossLabelFormat.str());
-    }
-  }
+OsmReadSession::OsmReadSession(tapeserver::drive::DriveInterface &drive,
+  const tapeserver::daemon::VolumeInfo &volInfo, const bool useLbp)
+  : ReadSession(drive, volInfo, useLbp) {
+  throw NotImplemented("Constructor not implemented");
 }
 
 }  // namespace tapeFile